requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
pandas==2.1.3
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6

# Optional: For Supabase integration
# supabase==2.0.2

# For development
pytest==7.4.3
httpx==0.25.2